loud_sound: Adjust HTTP/REsponse and Access logs with trace ID
The name of the traceID is "X-B3-TraceID". This value is output
in the app logs and now the Access Logs, so the correlation works
between the traces and access logs. The access logs include the
latency information of the request and the body processing.

The Access logs now include all the information needed for validating
the application, creating alerts, etc.