:new: add kubernetes readiness probe: verify cache
Kubernetes probes is supported by SpringBoot as described at
https://www.baeldung.com/spring-liveness-readiness-probes#actuator-probes

Added a new package "autoconfig" for updating the properties
of the service. Included the classes for supercash default
and kubernetes properties so we can always reuse them.

The class HealthProbeVerifier will postconstruct and will make
the service unavailable. Upon running at the specified rate,
it will update the availability in case the cache was built.
Since the supercash sales are used in the flow, this is used
as the readiness probe.

All the other changes were done based on the rename of the
properties to the new package.