:sparkles: Add ParkingTicketStateTransition capability
Adding the base model for the state transition. At a given time of status query we set this value
Refactored the methods from other services into this class.
Finally, adding repository to store the values.

The parkinglotTicket now is composed of the state transitions
and so it can be displayed for the user.