:zap: :bug: Performance fix for query ticket status from storage
The query executed before was not joining the tables of tickets, payments and states.

After adding the graph support, it uses OUTER JOIN properly, in addition to changing from List to Set, add fixes to Hashcode Equals to Payments, etc.

Adding EntityGraph to fetch all properties https://blog.ippon.tech/boost-the-performance-of-your-spring-data-jpa-application/#method1retrievingandloadingobjectswithquery

https://vladmihalcea.com/jpql-distinct-jpa-hibernate/
https://vladmihalcea.com/hibernate-multiplebagfetchexception/