:bug: Fix the Feign HTTP Client Logging Level from config
Finally found the issue. The code change from April 2021 hard-coded the value
for the log level to BASIC. The level should be FULL https://www.baeldung.com/java-feign-logging#logging-configuration

Moved a couple of properties added into the SERVER PROPERTIES into the updated CLIENT PROPERTIES.

Also, updating the corresponding application properties files and all the classes
that use the properties from the client, including the Application, that sets the level and must be according to the client.